수강 클릭시 정보 확인 진행중

+++ src/main/java/egovframework/veterans/com/cmm/service/vo/fmsc_s01.java
@@ -38,19 +38,19 @@ public class fmsc_s01{
 	private int DCID;
 
 
-	private double DiscountRate;
+	private int DiscountRate;
 
 
-	private double DCPrice;
+	private int DCPrice;
 
 
-	private double ItemPrice;
+	private int ItemPrice;
 
 
-	private double RealPrice;
+	private int RealPrice;
 
 
-	private double Misu;
+	private int Misu;
 
 
 	private String Note;
